golden-paths/h3-innovation/foundry-agent/skeleton/requirements.txt
azure-identity>=1.15.0
azure-ai-projects>=1.0.0b1
openai>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
structlog>=24.0.0
